+++ paperboy/src/main/kotlin/com/github/porokoro/paperboy/ElementTypes.kt
@@ -1,0 +1,24 @@
+/*
+ * Copyright (C) 2015-2016 Dominik Hibbeln
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.porokoro.paperboy
+
+internal object ElementTypes {
+    val SECTION_HEADER = 1
+    val TYPE_HEADER = 2
+    val ITEM_NONE = 3
+    val ITEM_LABEL = 4
+    val ITEM_ICON = 5
+}