+++ paperboy/src/main/kotlin/com/github/porokoro/paperboy/PaperboyConfiguration.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright (C) 2015-2016 Dominik Hibbeln
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.porokoro.paperboy
+
+import android.util.SparseArray
+
+internal data class PaperboyConfiguration(
+        var file: String? = "",
+        var fileRes: Int = 0,
+        var viewType: Int = ViewTypes.NONE,
+        var sectionLayout: Int = 0,
+        var typeLayout: Int = 0,
+        var itemLayout: Int = 0,
+        var sortItems: Boolean = false,
+        var itemTypes: SparseArray<ItemType> = SparseArray())