+++ paperboy/src/main/kotlin/com/github/porokoro/paperboy/TouchDispatchingTextView.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright (C) 2015-2016 Dominik Hibbeln
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.porokoro.paperboy
+
+import android.content.Context
+import android.support.v7.widget.AppCompatTextView
+import android.util.AttributeSet
+import android.view.MotionEvent
+import android.view.ViewGroup
+
+/**
+ * A [AppCompatTextView] that dispatches its touch events to its parent if it's clickable. The touch event is
+ * dispatched to support clickable [AppCompatTextView]s and [ViewGroup]s with animated backgrounds at the
+ * same time.
+ */
+class TouchDispatchingTextView : AppCompatTextView {
+    constructor(context: Context?) : super(context)
+
+    constructor(context: Context?, attrs: AttributeSet?) : super(context, attrs)
+
+    constructor(context: Context?, attrs: AttributeSet?, defStyleAttr: Int) : super(context, attrs, defStyleAttr)
+
+    override fun onTouchEvent(event: MotionEvent?): Boolean {
+        if (isClickable && parent != null && parent is ViewGroup)
+            (parent as ViewGroup).onTouchEvent(event)
+
+        return super.onTouchEvent(event)
+    }
+}