+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Version 3.0.0
+-------------
+- Complete conversion to the Kotlin language.
+- Android Support Libraries updated to version 23.2.1.
+
 Version 2.1.0
 -------------
 - Android Support Libraries updated to version 23.1.0.